removed 'Session' object from schema and codebase in general.
