+++ src/main/java/com/deco/tickexperience/model/entity/Event.java
@@ -1,15 +1,11 @@
 package com.deco.tickexperience.model.entity;
 
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.persistence.*;
 import lombok.Data;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Entity
 @Data
@@ -18,15 +14,26 @@ public class Event {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String name;
-    private String location;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
 
     @ManyToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name="SessionId")
-    private Session session;
-
-    @ManyToOne(cascade = CascadeType.ALL)
     @JoinColumn(name="SportId")
     private Sport sport;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name="VenueId")
+    private Venue venue;
+
+    @OneToMany(mappedBy = "event", cascade = CascadeType.ALL)
+    @JsonIgnore
+    private List<Ticket> tickets;
+
+    public void addTicket(Ticket ticket) {
+        ticket.setEvent(this);
+        if (this.getTickets() == null) {
+            this.setTickets(new java.util.ArrayList<>());
+        }
+        this.getTickets().add(ticket);
+    }
 }
