+++ src/main/java/com/deco/tickexperience/repository/TicketRepository.java
@@ -1,0 +1,7 @@
+package com.deco.tickexperience.repository;
+
+import com.deco.tickexperience.model.entity.Ticket;
+import org.springframework.data.repository.CrudRepository;
+
+public interface TicketRepository extends CrudRepository<Ticket, Long> {
+}
