+++ src/main/java/com/deco/tickexperience/repository/VenueRepository.java
@@ -1,0 +1,7 @@
+package com.deco.tickexperience.repository;
+
+import com.deco.tickexperience.model.entity.Venue;
+import org.springframework.data.repository.CrudRepository;
+
+public interface VenueRepository extends CrudRepository<Venue, Long> {
+}
