+++ src/main/resources/db/changelog/h2_setup.sql
@@ -1,10 +1,3 @@
-CREATE TABLE Session (
-    Id INT PRIMARY KEY AUTO_INCREMENT NOT NULL,
-    Location VARCHAR(255) NOT NULL,
-    StartTime TIMESTAMP NOT NULL,
-    EndTime TIMESTAMP NOT NULL
-);
-
 CREATE TABLE Sport (
     Id INT PRIMARY KEY AUTO_INCREMENT NOT NULL,
     Name VARCHAR(255) NOT NULL
@@ -12,14 +5,13 @@ CREATE TABLE Sport (
 
 CREATE Table Event (
     Id INT PRIMARY KEY AUTO_INCREMENT NOT NULL,
-    SessionId INT NOT NULL,
     SportId INT NOT NULL,
+    VenueId INT NOT NULL,
     Name VARCHAR(255) NOT NULL,
-    Location VARCHAR(255) NOT NULL,
     StartTime TIMESTAMP NOT NULL,
     EndTime TIMESTAMP NOT NULL,
-    FOREIGN KEY (SessionId) REFERENCES Session(Id),
-    FOREIGN KEY (SportId) REFERENCES Sport(Id)
+    FOREIGN KEY (SportId) REFERENCES Sport(Id),
+    FOREIGN KEY (VenueId) REFERENCES Venue(Id)
 );
 
 CREATE TABLE User (
@@ -35,11 +27,16 @@ CREATE TABLE User (
 
 CREATE TABLE Ticket (
     Id INT PRIMARY KEY AUTO_INCREMENT NOT NULL,
-    UserId INT,
-    SessionId INT NOT NULL,
+    UserId INT NOT NULL,
+    EventId INT NOT NULL,
     SeatNumber INT NOT NULL,
-    Venue VARCHAR(255) NOT NULL,
     FOREIGN KEY (UserId) REFERENCES User(Id),
-    FOREIGN KEY (SessionId) REFERENCES Session(Id),
-    UNIQUE (SeatNumber, Venue)
+    FOREIGN KEY (EventId) REFERENCES Event(Id),
+    UNIQUE (SeatNumber)
 );
+
+CREATE TABLE Venue (
+    Id INT PRIMARY KEY AUTO_INCREMENT NOT NULL,
+    Name VARCHAR(255) NOT NULL,
+    Location VARCHAR(255) NOT NULL,
+)
